Cache mcal
Cache mcal ; Move dats into separate file ; Fix read-from-cache if not exist
