+++ src/yfinance_cache/yfc_cache_manager.py
@@ -1,20 +1,19 @@
 import os
 import pickle, json
+from pprint import pprint
 
 from datetime import datetime, timedelta
 from zoneinfo import ZoneInfo
 
-# from .yfc_utils import *
+from yfc_dat import *
 from yfc_utils import *
-
-import yfc_time
 
 # To reduce #files in cache, store some YF objects together into same file (including metadata)
 packed_data_cats = {}
 # packed_data_cats["info"] = ["info"]
 packed_data_cats["quarterlys"] = ["quarterly_balance_sheet", "quarterly_cashflow", "quarterly_earnings", "quarterly_financials"]
 packed_data_cats["annuals"]    = ["balance_sheet", "cashflow", "earnings", "financials"]
-for i in yfc_time.intervalToString.values():
+for i in intervalToString.values():
 	packed_data_cats["history-"+i] = ["history-"+i]
 
 quarterly_objects = packed_data_cats["quarterlys"]
@@ -95,15 +94,19 @@ def ReadCacheDatum(ticker, objectName):
 	if os.path.isfile(fp_base+".json") and os.path.isfile(fp_base+".pkl"):
 		raise Exception("For cached datum '{0}', both a .json and .pkl file exist. Should only be one.".format(objectName))
 
+	exists = False
 	if os.path.isfile(fp_base+".json"):
 		fp = fp_base+".json"
-		with open(fp, 'r') as inData:
-			js   = json.load(inData, object_hook=JsonDecodeDict)
-			data = js["data"]
-			md   = js["metadata"]
+		if os.path.isfile(fp):
+			exists = True
+			with open(fp, 'r') as inData:
+				js   = json.load(inData, object_hook=JsonDecodeDict)
+				data = js["data"]
+				md   = js["metadata"]
 	else:
 		fp = fp_base+".pkl"
 		if os.path.isfile(fp):
+			exists = True
 			with open(fp, 'rb') as inData:
 				pkl = pickle.load(inData)
 				if isinstance(pkl, dict):
@@ -115,6 +118,11 @@ def ReadCacheDatum(ticker, objectName):
 				else:
 					raise Exception("Pickle missing metadata: "+fp)
 					# data = pkl
+	if not exists:
+		return None
+
+	if md is None:
+		raise Exception("Failed to find metadata in file: "+fp)
 
 	if "Expiry" in md.keys():
 		dt = datetime.utcnow().replace(tzinfo=ZoneInfo("UTC"))
@@ -163,6 +171,9 @@ def StoreCacheDatum(ticker, objectName, datum, expiry=None):
 	if verbose:
 		print("StoreCacheDatum({0}, {1})".format(ticker, objectName))
 
+	if isinstance(expiry, Interval):
+		expiry = datetime.utcnow().replace(tzinfo=ZoneInfo("UTC")) + intervalToTimedelta[expiry]
+
 	if IsObjectInPackedData(objectName):
 		StoreCachePackedDatum(ticker, objectName, datum, expiry)
 		return
@@ -186,7 +197,12 @@ def StoreCacheDatum(ticker, objectName, datum, expiry=None):
 
 	fp = fp_base+"."+ext
 
+	if verbose:
+		print("- storing {} as {} at {}".format(objectName, ext, fp))
+
 	if not os.path.isfile(fp):
+		if verbose:
+			print("- doesn't exist: "+fp)
 		md = {}
 	else:
 		if ext == "json":
